Windows was complaining about size_t/long loss of data, type casted all items and warnings are gone. Changed the forloop to a BOOST_FOREACH loop.

+++ waveq3d/wave_queue.cc
@@ -1020,7 +1020,7 @@ bool wave_queue::checkEigenrayListeners(long waveTime){
     for (std::vector<eigenrayListener*>::iterator iter = _eigenrayListenerVec.begin();
                                                 iter != _eigenrayListenerVec.end(); ++iter){
         eigenrayListener* pListener = *iter;
-        pListener->checkEigenrays(_run_id, waveTime);
+        pListener->checkEigenrays((unsigned long)_run_id, waveTime);
     }
 
     return (_eigenrayListenerVec.size() > 0);
